fix(impress): Add icons for new exports
The new exports were added in I17eee20c78491f3f6a2a7e24293050e0821a0659.

Most of these are "just" images, so (although not ideal) I'll just use
the same icon we do for png. I've done something different for svg, as
it's a vector, and something different for shockwave flash.

I have included dark formats for the icons (as I had believed icons
would be specified in a different way so the dark formats would be
used), however the notebookbar css doesn't yet do anything with them. In
the future, we should change this so that it does.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I17eee20c78491f3f6a2a7e24293050e0821a0659