Remove blocking function from the online - 1. step
The blocking for functions was introduced in July 2021 with commit
https://github.com/CollaboraOnline/online/commit/e02e1df76d8379e4da7e7c87f10c594bc558d6d2
("Introduce Freemium options") and after some additional edits renamed /
refactored to locking later to locking
 (Change-Id:I2935edbfe1eb8dcc194641afa2461cf77b2b29d3).

There is no need for a function that holds back functions from users
at compilation time. And it is not in accordance with the principles
of Free Software under a copyleft license from my view.

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: I3a21e07a50e2457e3f221aa870e856161153331b