slideshow: support effects based on transition filters
This patch refactors transition implementation in order to enable support for
entrance and exit effects based on transition filters.
The involved transition filters are the ones based on clipping/masking.
Now ClippingTransition handles 'mode out' and reverse direction when reverse
method is of type SubtractAndInvert. That simplified Box and Split mask
function.
ClippingTransition handles isotropic scaling, too. This affects circle, dissolve
and wheel transitions.
The 'Plus' transition now handles the cornersout subtype which is the one used
for slide transitions.
The ellipsewipe transition, now uses a real circle as mask for the circle
subtype used in shape/circle slide transition.
Now slidewipe is mapped to barwipe as on desktop for shape transitions. See
createShapeTransition in AnimationFactory.ts. That allows the 'Peek in' effect
to work as expected.

New file: TransitionInfo.ts. It contains parameters needed in order to implement
fragment shaders and mask functions for transition filters.

Adapt patch to LibreOffice Online.

Signed-off-by: Marco Cecchetti <marco.cecchetti@collabora.com>
Change-Id: Ia2bd36cfbc741d64a1f92f08375f4cc1668787c8