Option to hide server signature in HTTP response
Hiding the server signature reduces the likelihood of targeted
attacks by making it harder for attackers to identify the
server's configuration.

In built-in default configuration security.server_signature is
set to false. It is also set to false in loolwsd.xml packaged
with production builds. HTTP response header is like:
    Server: LOOLWSD HTTP Server
In debug mode security.server_signature is set to true in
coolwsd.xml. HTTP response header is like:
    Server: LOOLWSD HTTP Server 24.04.9.1

Adapt patch to LibreOffice Online.

Signed-off-by: Andras Timar <andras.timar@collabora.com>
Change-Id: Ie23c2dbc1514e46e2942f558ada2e27d0f108bea