Make tile decompression transactional
Instead of sending compressed updates directly to the Worker on-demand,
this groups updates in transactions around the event blocks established
in socket.js and elsewhere, and disallows updates while these
transactions are being processed.

The intention is that the worker path ought to work in roughly the same
way as the blocking non-worker path, except that the UI won't be blocked
while updates are being processed.

Adapt patch to LibreOffice Online.

Signed-off-by: Chris Lord <chris.lord@collabora.com>
Change-Id: I86dd339dec4cadc1fca9c0bcd220099742e653d7