wsd: use the 'flush' logging property to disable buffering
In some cases it's benefitial to disable buffered-
logging and flush after each entry is generated.

We already have an unbuffered logger and a
property, called 'flush', already exists and
is used by Poco to also disable buffering its
file-logger.

We leverage this property to disable buffering
in both the console logger (this patch), as well
as in Poco's file-logger (existing functionality).

Change-Id: I5e3b75e66f82a8a5df57901ef4f6c8264aee3565
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>