feat(touch): Determine last-used input method
Previously, when we wanted to determine if some touch UI should be shown
we would check if there were any touchscreens on the system. This is
still used in some places, but it turns out that this is poor in many
cases (a key example would be the cursor droplet, which showed on
touchscreen laptops even when the touchscreen was not being used, for
example).

You might think that this problem is easily solved: "Just check if the
thing that clicked into the document was a touchscreen, and only show
the droplet then". We can't, however, trivially detect if the input that
caused the cursor to show was from a touchscreen as we roundtrip through
core before showing the cursor. This sort of property (hard to get a
triggering event) holds true for the other events we were using
hasAnyTouchscreen for previously.

Instead, we can listen to *all* mouse or tap events, and change our
behavior based on whatever was triggered last. This isn't quite the
"triggering event" behavior (e.g. if we close a jsdialog with a
touchscreen we will get a droplet cursor, when arguably we might not
want that) but it'll be the same a vast majority of the time (and I am
undecided on whether this behavior is in-fact preferable in the cases
where it differs)

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I88f7936495415ad04012fa998b1bd242b3bd8f5b