Status bar: fix dropdown width
Their container is now a flex box so we should allow them to grow but not to
shrink but that would not solve the problem when there is no space to
fit content (example higher UI zoom). The best would be to set the
button in a way that doesn't allow in any circumstances wrapping /
cropping.

Signed-off-by: Pedro Pinto Silva <pedro.silva@collabora.com>
Change-Id: Ide87f81d403e1f77977c6098d6f1a6e85828468e