cool#9992 doc electronic sign: add menu item to trigger hash extract
In short, the trouble with signing with PEM files is that many CAs won't
hand out software certs for signing, need to use a 3rd-party to do the
signing. This requires hash-extract / upload / sign / download /
serialize pipeline on our side.

This commit starts with the first step, to extract a hash of the
document to be signed.

Given that the hash depends on the time (the signed document has a
signature widget, and that contains a timestamp), also get that from
core. Do this with a new menu item of the Draw / PDF UI: only visible
when the user private info from WOPI has a ESignatureBaseUrl key, which
will be the API server we'll talk to in the future. (So all this is
hidden by default, allowing this add this incrementally.)

Also fix a missing make dependency, so COOL_JS.m4 gets rebuilt on adding
a new .ts file, to avoid the need to do a 'make clean' on adding a new
file, which sounded odd.

Adapt patch to LibreOffice Online.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: Ib9f15fc1448eea4a4fa1784cf0d9b3de00c44019