Introduce CSS color var doc-type and remove color dependent rules
We now have data-doctype attribute that is being added to the <body>
as an additional attribute so, use that to set a single css color var
that changes depending on the attributes value and remove duplicated
css bits.

Additional note: for now it seems we can just use those colors also
for dark mode so, no need to add additional lines to the color-palette-dark.css

Adapt patch to LibreOffice Online.

Signed-off-by: Pedro Pinto Silva <pedro.silva@collabora.com>
Change-Id: Ife75487b3dfbe5d6ec6e4c961721c8734520786e