cool#9992 doc electronic sign: select the provider interactively
This allows not storing the name of the provider in the user private
info, which had two issues: the list of available providers is only
known after sending the hash & the provider ID is not supposed to be
visible to the user (all providers have a human-readable name).

This requires splitting handleSendHashJson() to two: once we sent the
hash, we can show the dialog, then the dialog calls back to the new
handleSelectedProvider().

The details about providers is available with an API, but similar to
Zotero, it would be pointless to re-fetch these every time for every
user, so just have them inline (the test env would take about 0.5 sec
each time to fetch this list).

Update the cypress test accordingly, which points out that
eideasy-send-hash.json was wrong (it was the response to the "get
signature" API) and lacked an "available_methods" field.

Adapt patch to LibreOffice Online.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: I88b3e8180c692ac63bf48b53dcbf17bf299618b2