cool#10630 doc electronic sign: implement support for redirect-based providers
Some eideasy providers are in-context (e.g. smart-id-signature) and
those were already working, some others are redirect-based (e.g.
d-trust-sign-me-qes-signature).

Redirect-based methods need a URL they can visit after signing, set this
to the recently introduced /cool/signature endpoint, which will do the
same postMessage() as the stock (external) in-context esign page does.

Also log the selected provider ID, that is potentially useful if later
people have problem with a specific provider.

Finally set an option, to avoid the unwanted default behavior that the
created PKCS#7 blob would be downloaded in the browser -- this makes
sense for entire PDF files, but not for signature binaries.

Adapt patch to LibreOffice Online.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: Idd9a17b962c0c0951618b6b94091dbe2fb45bd80