feat(touch): add touch detection for mouseover
On some devices, such as iPads, mouseover events are emulated when an
element is tapped. They don't provide information on which device used
them and are of the type 'MouseEvent' so we can't detect whether they're
actually triggered by a touch from that.

However, if mouseover events occur immediately after a touch we can be
somewhat sure that it's really a touch event. Similarly, if a mouseover
event occurs without a touch we can be somewhat sure that it's just
someone moving a mouse. This patch exploits that to extend our
window.touch.isTouchEvent function to handle these event types.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ie42d1718c123796ed330df7cedd3af7fa1f4ce78